refactor alcohol 조회, 찜하기

+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/storage/BeerStorage.java
@@ -17,7 +17,7 @@ public class BeerStorage {
 	
 	int highestPrice;
 	
-	double abv;
+	float abv;
 	
 	String kind;
