+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/storage/LiquorStorage.java
@@ -17,7 +17,7 @@ public class LiquorStorage {
 	
 	int highestPrice;
 	
-	double abv;
+	float abv;
 	
 	String kind;
