+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/storage/WineStorage.java
@@ -17,7 +17,7 @@ public class WineStorage {
 	
 	int highestPrice;
 	
-	double abv;
+	float abv;
 	
 	String kind;
